requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
apache-airflow==3.0.4
apache-airflow-core==3.0.4
apache-airflow-providers-common-compat==1.7.3
apache-airflow-providers-common-io==1.6.2
apache-airflow-providers-common-sql==1.27.5
apache-airflow-providers-smtp==2.2.0
apache-airflow-providers-standard==1.6.0
apache-airflow-task-sdk==1.0.4
argcomplete==3.6.2
asgiref==3.9.1
asyncio==3.4.3
attrs==25.3.0
azure-core==1.35.0
azure-storage-blob==12.26.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
cadwyn==5.4.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.6
-e git+https://github.com/ProductPriceTrackerOrg/data-pipeline.git@b91ba4a0f0ba46b5acd093c4f5c78c26470df76d#egg=data_pipeline
DateTime==5.5
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
Flask==3.1.2
frozenlist==1.7.0
fsspec==2025.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-bigquery==3.37.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.12.0
libcst==1.8.2
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==6.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
msgspec==0.19.0
multidict==6.6.3
nest-asyncio==1.5.6
numpy==2.3.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
retryhttp==1.3.3
rich==14.1.0
rich-argparse==1.7.1
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.27.0
rsa==4.9.1
selectolax==0.3.17
sentry-sdk==2.35.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.47.2
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
typer==0.16.1
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
